Remove pointless cast

+++ src/Controllers/PostController.php
@@ -110,7 +110,7 @@ class PostController
         return $this->renderer->render($response, 'post-list', [
             'posts' => $parsedPosts,
             'current_page' => $page,
-            'total_pages' => (int) ceil($total / $limit) ?: 1,
+            'total_pages' => ceil($total / $limit) ?: 1,
         ]);
     }
